CX/httpfunc/py/requirements.txt
flask==1.1.2
google-cloud-error-reporting==1.1.2
google-cloud-datastore==2.1.0